hw4 auth implementation

+++ controllers/contacts/services.js
@@ -1,18 +1,19 @@
 const Contact = require("../../models/contacts");
 
-const fetchContacts = () => {
-  return Contact.getAll();
+const fetchContacts = (userId) => {
+  return Contact.find({ owner: userId });
 };
 
-const fetchContact = (id) => {
-  return Contact.findById({ _id: id });
+const fetchContact = ({ id, userId }) => {
+  return Contact.findById({ _id: id, owner: userId });
 };
 
-const addContact = ({ name, email, phone }) => {
-  return Contact.create({ name, email, phone });
+const addContact = ({ name, email, phone, owner }) => {
+  return Contact.create({ name, email, phone, owner });
 };
-const removeContact = (id) => {
-  return Contact.deleteOne({ _id: id });
+
+const removeContact = ({ id, owner }) => {
+  return Contact.deleteOne({ _id: id, owner: owner });
 };
 
 const updateContact = ({ id, toUpdate, upsert = false }) => {
